% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{best_lambda}
\alias{best_lambda}
\title{Extract best lambda from CV object}
\usage{
best_lambda(model, which = c("min", "1se"))
}
\arguments{
\item{model}{cv.rvflnet object}

\item{which}{which lambda to return ("min" or "1se")}
}
\value{
numeric lambda value
}
\description{
Extract best lambda from CV object
}
