% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{coef.cv.rvflnet}
\alias{coef.cv.rvflnet}
\title{Extract coefficients from cross-validated RVFLNet model}
\usage{
\method{coef}{cv.rvflnet}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{cv.rvflnet object}

\item{s}{lambda value ("lambda.min", "lambda.1se", or numeric)}

\item{...}{additional arguments passed to coef.glmnet}
}
\value{
coefficient matrix. See coef.rvflnet for interpretation notes.
}
\description{
Extract coefficients from cross-validated RVFLNet model
}
\examples{
\dontrun{
coef(cv_model, s = "lambda.min")
}
}
