% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{cv.rvflnet}
\alias{cv.rvflnet}
\title{Cross-validated RVFLNet model}
\usage{
cv.rvflnet(
  x,
  y,
  n_hidden = 200L,
  activation = c("sigmoid", "tanh", "relu", "identity"),
  W_type = c("gaussian", "uniform", "sobol"),
  seed = 1,
  scale = TRUE,
  include_original = TRUE,
  store_y = FALSE,
  family = c("gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"),
  ...
)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{n_hidden}{number of hidden units}

\item{activation}{activation function}

\item{W_type}{random feature type ("gaussian", "uniform", "sobol")}

\item{seed}{random seed}

\item{scale}{logical scaling}

\item{include_original}{logical, whether to include original features}

\item{store_y}{logical, whether to store y in model}

\item{family}{response type}

\item{...}{additional arguments passed to glmnet::cv.glmnet}
}
\value{
object of class "cv.rvflnet"
}
\description{
Fits RVFL features then applies cv.glmnet.
}
\examples{
\dontrun{
x <- matrix(rnorm(100*5), 100, 5)
y <- x[,1] + sin(x[,2]) + rnorm(100, 0, 0.1)
cv_model <- cv.rvflnet(x, y, n_hidden = 50, nfolds = 5)
plot(cv_model)
predict(cv_model, newx = x[1:10,])
}
}
