% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{.apply_scaling}
\alias{.apply_scaling}
\title{Apply scaling to new data}
\usage{
.apply_scaling(x, center, scale_vec)
}
\arguments{
\item{x}{input matrix}

\item{center}{centering vector}

\item{scale_vec}{scaling vector}
}
\value{
scaled matrix
}
\description{
Apply scaling to new data
}
