% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{.compute_scaling}
\alias{.compute_scaling}
\title{Compute scaling parameters from data}
\usage{
.compute_scaling(x, scale_input = TRUE)
}
\arguments{
\item{x}{input matrix}

\item{scale_input}{logical, whether to scale}
}
\value{
list with center and scale vectors
}
\description{
Compute scaling parameters from data
}
