% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{.rvfl_features}
\alias{.rvfl_features}
\title{Generate RVFL random features}
\usage{
.rvfl_features(x, W, center, scale_vec, activation)
}
\arguments{
\item{x}{numeric matrix of predictors}

\item{W}{random projection matrix}

\item{center}{centering vector}

\item{scale_vec}{scaling vector}

\item{activation}{activation function name or custom function}
}
\value{
augmented design matrix cbind(x, hidden features)
}
\description{
Internal helper to build the RVFL design matrix.
}
