% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{fitted.cv.rvflnet}
\alias{fitted.cv.rvflnet}
\title{Extract fitted values from cross-validated RVFLNet model}
\usage{
\method{fitted}{cv.rvflnet}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{cv.rvflnet object}

\item{s}{lambda value}

\item{...}{additional arguments}
}
\value{
fitted values
}
\description{
Extract fitted values from cross-validated RVFLNet model
}
