% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{fitted.rvflnet}
\alias{fitted.rvflnet}
\title{Extract fitted values from RVFLNet model}
\usage{
\method{fitted}{rvflnet}(object, s = NULL, ...)
}
\arguments{
\item{object}{rvflnet object}

\item{s}{lambda value}

\item{...}{additional arguments}
}
\value{
fitted values
}
\description{
Extract fitted values from RVFLNet model
}
