% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{get_activation}
\alias{get_activation}
\title{Extract activation function}
\usage{
get_activation(model)
}
\arguments{
\item{model}{rvflnet or cv.rvflnet object}
}
\value{
character
}
\description{
Extract activation function
}
