% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{get_weight_type}
\alias{get_weight_type}
\title{Extract weight type}
\usage{
get_weight_type(model)
}
\arguments{
\item{model}{rvflnet or cv.rvflnet object}
}
\value{
character
}
\description{
Extract weight type
}
