% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{is.cv.rvflnet}
\alias{is.cv.rvflnet}
\title{Check if object is a CV-RVFLNet model}
\usage{
is.cv.rvflnet(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Check if object is a CV-RVFLNet model
}
