% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{is.rvflnet}
\alias{is.rvflnet}
\title{Check if object is an RVFLNet model}
\usage{
is.rvflnet(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Check if object is an RVFLNet model
}
