% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{n_hidden}
\alias{n_hidden}
\title{Extract number of hidden units}
\usage{
n_hidden(model)
}
\arguments{
\item{model}{rvflnet or cv.rvflnet object}
}
\value{
integer
}
\description{
Extract number of hidden units
}
