% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{plot.cv.rvflnet}
\alias{plot.cv.rvflnet}
\title{Plot CV curve for RVFLNet}
\usage{
\method{plot}{cv.rvflnet}(x, ...)
}
\arguments{
\item{x}{cv.rvflnet object}

\item{...}{passed to plot.glmnet}
}
\description{
Plot CV curve for RVFLNet
}
\examples{
\dontrun{
plot(cv_model)
}
}
