% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{plot.rvflnet}
\alias{plot.rvflnet}
\title{Plot RVFLNet model}
\usage{
\method{plot}{rvflnet}(x, type = c("coef"), ...)
}
\arguments{
\item{x}{rvflnet object}

\item{type}{type of plot ("coef" for coefficient path)}

\item{...}{additional arguments passed to plot functions}
}
\description{
Generic plot method for rvflnet objects
}
