% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{print.cv.rvflnet}
\alias{print.cv.rvflnet}
\title{Print cross-validated RVFLNet model}
\usage{
\method{print}{cv.rvflnet}(x, ...)
}
\arguments{
\item{x}{cv.rvflnet object}

\item{...}{ignored}
}
\description{
Print cross-validated RVFLNet model
}
