% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{residuals.cv.rvflnet}
\alias{residuals.cv.rvflnet}
\title{Residuals method for cross-validated RVFLNet}
\usage{
\method{residuals}{cv.rvflnet}(object, s = "lambda.min", type = "response", ...)
}
\arguments{
\item{object}{cv.rvflnet object}

\item{s}{lambda value ("lambda.min", "lambda.1se", or numeric)}

\item{type}{type of residuals}

\item{...}{additional arguments}
}
\value{
residuals vector
}
\description{
Residuals method for cross-validated RVFLNet
}
