% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{residuals.rvflnet}
\alias{residuals.rvflnet}
\title{Residuals method for RVFLNet}
\usage{
\method{residuals}{rvflnet}(object, s = NULL, type = "response", ...)
}
\arguments{
\item{object}{rvflnet object}

\item{s}{lambda value}

\item{type}{type of residuals ("response", "deviance", etc.)}

\item{...}{additional arguments}
}
\value{
residuals vector
}
\description{
Residuals method for RVFLNet
}
\examples{
\dontrun{
model <- rvflnet(x, y, store_y = TRUE)
residuals(model, s = "lambda.min")
}
}
