% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{summary.cv.rvflnet}
\alias{summary.cv.rvflnet}
\title{Summary method for cross-validated RVFLNet}
\usage{
\method{summary}{cv.rvflnet}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{cv.rvflnet object}

\item{s}{lambda value ("lambda.min" or "lambda.1se")}

\item{...}{additional arguments}
}
\value{
invisible summary object
}
\description{
Summary method for cross-validated RVFLNet
}
