% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvflnet.R
\name{summary.rvflnet}
\alias{summary.rvflnet}
\title{Summary method for RVFLNet}
\usage{
\method{summary}{rvflnet}(object, s = NULL, ...)
}
\arguments{
\item{object}{rvflnet object}

\item{s}{lambda value}

\item{...}{additional arguments}
}
\value{
invisible summary object
}
\description{
Summary method for RVFLNet
}
